pyproject.toml
[tool.poetry]
name = "currency-bot-project"
version = "0.1.0"
description = ""
authors = ["Dmitriy <dimnissv@yandex.kz>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.14,<3.15"
python-dotenv = "^1.2.2"
aiofiles = "^25.1.0"
aiohttp = "^3.13.5"
aiogram = "^3.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
